Image and border are now effects
There is still much work to be done, but the basic framework is there.

After implementing the original effects system (basically layers), I
realized that it would also make sense to have both the images and the
borders also a part of that system, as it would allow users to place
multiple images and customize the border to their liking (or do CRAZY
things like have multiple or stacked borders).

Still existing is much old cruft that shall be cleaned out in due time.